reflect XML parsing more directly in classes
- `<PARAMETERS>` is parsed in new `Parameters`
- `<NODE>` is parsed in `ParameterGroup`
- `<ITEM>` and <ITEMLIST> is parsed in `Parameter`

This was mainly done to be able to use this module
for parsing paramXML (e.g. the ini files generated
by OpenMS).

In order to deal with

```
<PARAMETERS>
  <NODE name="TOOLNAME" ...>
    <ITEM name="version" />
    <NODE name="1" ...>
...
    </NODE>
  </NODE>
</PARAMETER>
```

parsing in `Parameters` now simply swallows the NODEs and ITEMs
shown in the above code (i.e. they are not represented in object
generated structure -- but the infos like version and despription
are saved).
The structure is restored when XML is generated using the
`_xml_node` function.

In addition:

- implement ModelTypeError that is raised for Parameters() and CTDNode
  if they have the wrong root tag (there was a TODO)
- use _Null as Singleton object: advantage is that a str() method can be
  used